r24317@hlagh (orig r9367):  tonvoon | 2010-05-14 12:24:35 -0400
Branch for converting eval {} to Try::Tiny

r24319@hlagh (orig r9369):  tonvoon | 2010-05-14 17:25:02 -0400
Conversion of eval => try (part 1)

r24325@hlagh (orig r9375):  tonvoon | 2010-05-14 18:03:03 -0400
Add eval => try

r24326@hlagh (orig r9376):  tonvoon | 2010-05-14 18:22:57 -0400
Another eval => try

r24327@hlagh (orig r9377):  tonvoon | 2010-05-14 18:45:27 -0400
Corrected usage of $@ in catch block

r24328@hlagh (orig r9378):  tonvoon | 2010-05-14 19:29:52 -0400
txn_do's eval => try

r24329@hlagh (orig r9379):  tonvoon | 2010-05-14 19:46:44 -0400
eval => try where tests for $@ done

r24330@hlagh (orig r9380):  tonvoon | 2010-05-14 20:38:43 -0400
All expected evals converted to try, except where no test is done,
runtime evaluation, or base perl (such as "require"). Only one test
failure due to string difference in output

r24346@hlagh (orig r9396):  tonvoon | 2010-05-17 08:52:28 -0400
Fix missing $@ in try::tiny conversion

r24347@hlagh (orig r9397):  tonvoon | 2010-05-17 08:55:13 -0400
Revert to eval instead of try::tiny because no check for $@

r24348@hlagh (orig r9398):  tonvoon | 2010-05-17 08:55:45 -0400
Added myself to contributors

r24349@hlagh (orig r9399):  tonvoon | 2010-05-17 10:23:57 -0400
Fixed exception logic due to not being able to use return with a catch{}

r24350@hlagh (orig r9400):  tonvoon | 2010-05-17 10:31:32 -0400
Removed tab

r24430@hlagh (orig r9424):  ribasushi | 2010-05-25 10:09:39 -0400
More try::tiny conversions
r24432@hlagh (orig r9426):  ribasushi | 2010-05-25 11:40:45 -0400
Try::Tiny conversion finished
r24433@hlagh (orig r9427):  ribasushi | 2010-05-25 11:46:52 -0400
Missed use
r24440@hlagh (orig r9434):  rkitover | 2010-05-25 13:47:25 -0400
fix Oracle
r24441@hlagh (orig r9435):  rkitover | 2010-05-25 14:04:10 -0400
fix odbc/mssql dynamic cursors
r24442@hlagh (orig r9436):  rkitover | 2010-05-25 14:32:41 -0400
fix hang in SQLAnywhere DateTime tests

+++ t/inflate/datetime_sybase_asa.t
@@ -3,6 +3,7 @@ use warnings;
 
 use Test::More;
 use Test::Exception;
+use Scope::Guard ();
 use lib qw(t/lib);
 use DBICTest;
 
@@ -27,20 +28,20 @@ my @info = (
   [ $dsn2, $user2, $pass2 ],
 );
 
-my @handles_to_clean;
+my $schema;
 
 foreach my $info (@info) {
   my ($dsn, $user, $pass) = @$info;
 
   next unless $dsn;
 
-  my $schema = DBICTest::Schema->clone;
+  $schema = DBICTest::Schema->clone;
 
   $schema->connection($dsn, $user, $pass, {
     on_connect_call => [ 'datetime_setup' ],
   });
 
-  push @handles_to_clean, $schema->storage->dbh;
+  my $sg = Scope::Guard->new(\&cleanup); 
 
 # coltype, col, date
   my @dt_types = (
@@ -72,7 +73,7 @@ SQL
       ->search({ trackid => $row->trackid }, { select => [$col] })
       ->first
     );
-    is( $row->$col, $dt, 'DateTime roundtrip' );
+    is( $row->$col, $dt, "$type roundtrip" );
 
     is $row->$col->nanosecond, $dt->nanosecond,
         'nanoseconds survived' if 0+$dt->nanosecond;
@@ -82,8 +83,8 @@ SQL
 done_testing;
 
 # clean up our mess
-END {
-  foreach my $dbh (@handles_to_clean) {
+sub cleanup {
+  if (my $dbh = $schema->storage->dbh) {
     eval { $dbh->do("DROP TABLE $_") } for qw/track/;
   }
 }
